requirements.txt
keras==2.12.0
matplotlib==3.7.1
matplotlib==3.6.3
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
tabulate==0.9.0
tensorflow==2.12.0
